refactor: storage object
