+++ src/objects/Playlist.ts
@@ -1,5 +1,4 @@
 import { v4 as uuidv4 } from 'uuid';
-import { dummyFavList } from '../utils/ChromeStorage';
 
 export * from '@APM/objects/Playlist';
 
@@ -11,4 +10,35 @@ export const defaultSearchList = ({
   newList.songList = songList;
   newList.info = info;
   return newList;
+};
+
+export const dummyFavList = (favName: string): NoxMedia.Playlist => {
+  return {
+    songList: [],
+    info: { title: favName, id: `FavList-${uuidv4()}` },
+    title: favName,
+    id: `FavList-${uuidv4()}`,
+    // this is not a Set because we need to serialize this
+    // for importing/exporting playlists.
+    subscribeUrls: [],
+    settings: {
+      autoRSSUpdate: false,
+    },
+    useBiliShazam: false,
+    biliSync: false,
+    bannedBVids: [],
+    showFavoriteList: false,
+  };
+};
+
+export const dummyFavListFromList = (
+  list: NoxMedia.Playlist,
+): NoxMedia.Playlist => {
+  return { ...dummyFavList(''), ...list };
+};
+
+export const dummyFavFavList = () => {
+  const favfavlist = dummyFavList('我的最爱');
+  favfavlist.info.id = 'FavList-Special-Favorite';
+  return favfavlist;
 };
